+TeamEntity => controller, services

+++ src/controllers/times.controller.js
@@ -27,68 +27,53 @@ const findAllTimesByPositionController = (req, res) => {
 };
 
 const createTimeController = (req, res) => {
-  const time = req.body;
-  if (
-    !time ||
-    !time.time ||
-    !time.time.nome_popular ||
-    !time.time.sigla ||
-    !time.time.escudo ||
-    !time.vitorias ||
-    !time.empates ||
-    !time.derrotas ||
-    !time.gols_pro ||
-    !time.gols_contra
-  )
+  try {
+    const time = new timesService.TeamEntity(req.body);
+    time.validateTeam();
+    const newTime = timesService.createTimeService(time);
+    res.send(newTime);
+  } catch (err) {
     return res.status(400).send({
-      message: "You didin't fill all the required data fields",
+      message: err.message,
     });
-
-  const newTime = timesService.createTimeService(time);
-  res.send(newTime);
+  }
 };
 
 const updateTimeController = (req, res) => {
-  const editedId = Number(req.params.id);
+  try {
+    const editedTime = req.body;
+    const editedId = Number(req.params.id);
+    if (!editedId || isNaN(editedId))
+      throw new Error('An ID is required for this request.');
+    editedTime.time.time_id = editedId;
+    const editedEntity = new timesService.TeamEntity(editedTime);
+    editedEntity.validateTeam();
+    const updatedTime = timesService.updateTimeService(editedEntity);
 
-  if (!editedId)
-    return res
-      .status(400)
-      .send({ message: 'An ID is required for this request.' });
-
-  const editedTime = req.body;
-
-  if (
-    !editedTime ||
-    !editedTime.time ||
-    !editedTime.time.nome_popular ||
-    !editedTime.time.sigla ||
-    !editedTime.time.escudo ||
-    !editedTime.vitorias ||
-    !editedTime.empates ||
-    !editedTime.derrotas ||
-    !editedTime.gols_pro ||
-    !editedTime.gols_contra
-  )
+    res.send(updatedTime);
+  } catch (err) {
     return res.status(400).send({
-      message: "You didin't fill all the required data fields",
+      message: err.message,
     });
-
-  const updatedTime = timesService.updateTimeService(editedId, editedTime);
-  res.send(updatedTime);
+  }
 };
 
 const deleteTimeController = (req, res) => {
-  const idParam = Number(req.params.id);
+  try {
+    const idParam = Number(req.params.id);
+    if (!idParam || isNaN(idParam))
+      throw new Error('An ID is required for this request.');
 
-  if (!idParam)
-    return res
-      .status(400)
-      .send({ message: 'An ID is required for this request.' });
+    const deletedTime = timesService.deleteTimeService(idParam);
 
-  const deletedTime = timesService.deleteTimeService(idParam);
-  res.send(deletedTime);
+    res.send(deletedTime);
+  } catch (err) {
+    return res.status(400).send({
+      message: err.message,
+    });
+  }
 };
+// END OF CONTROLLERS
 
 module.exports = {
   findAllTimesControler,
